10130 FIX Fixed performance bug in monitoring history cache
During startup, the CMC caches information from the monitoring history up to
a given horizon in the past.  This caching had a performance bug (quadratic
complexity), resulting in caching times of several days or even more,
depending on the content of the monitoring history.  This has been fixed,
caching time is now linear in the size of the monitoring history.

Change-Id: Id7e8ddc45500d4ec61d2ee4274889aae0a4ba742