Move PersistedSection filtering to PersistedSections
Now we have the following methods
 * SectionStore.load() loads the DB and returns its content.
 * SectionStore.store() updates the DB with new content or removes
   it if it is empty.
 * PersistedSections.filter() filters the persisted sections in place.

Moving the filtering from the SectionStore to the PersistedSections
reduces feature envy and lets the SectionStore class manage the DB
but ignore its content.

Note that the filtering criterion is not correct but this patch
does not change the behavior of the code but only its API.

We also add a few comments that explain the current code duplication
and provide a solution to get rid of it.

CMK-6516

Change-Id: Ic43bd104c08ea047cb42edb59db92c7bced4a454