Extract common to SNMP and Agent
This reduces the code duplication.

Note that the code should *not* be in HostSections.

Actually, most methods if not every method of HostSections should be in
the Parser and HostSections should only contain Final data.  This way,
the parsing would not be split between the parser and HostSections.
Then, we could instantiate HostSections with useful data instead of
whatever is currently available and modify the content of the class
before finally returning it.

This, however, requires larger changes to the parsers.

CMK-6516

Change-Id: Ia2d37d097773dd65817762e33315c31830d3527b