Inline `update_and_store`
The method was doing too much on the class: updating its argument and
the itself.  The algorithm is clearer inline.

It is also more flexible in that the `filter()` method with faulty logic
is now exposed to the caller and therefore easier to remove or fix.

Also rename `../base/../test_host_sections.py` to avoid name conflicts.

CMK-6516

Change-Id: I04559e32c21eabb325832766ccdb96d527eb2eec