HaSI: raise for path-reuse
Inventory plugins previously appeared tobe  able to create
Attributes and TableRows under the same path.
However, this lead to a crash when serializing the data.
We now raise an exception immediately when the offending
inventory element is processed. This way the exeption stays
in the context of the plugin.

CMK-6719

Change-Id: Iaf62e95261711b4126820d9153c593608e9344a3