11764 FIX Do not treat site as dead when EC is not running
The Event Console is used as data source for Event Console
related views. Since it is a dedicated process, it may be unreachable
while the monitoring core of a site is reachable.

Previous versions treated a site as down once the Event Console was not
running. This has now been changed. Sites in general are now still usable in
the GUI, even when the Event Console is stopped.

CMK-5540

Change-Id: Ib38cdc82bcaac7ef7ea24bb1ea226d7ab04c9d01