RTC: drop automation call
Since the latest refactorings in the agent, there are a lot more
sections, that can be run as real-time sections.

Either way, there is no point in attaching this information to a section
pl plugin, since whether any given "section" is a valid choice is only
known to the agent.

CMK-6742

Change-Id: Ia3727864ac80d8fa708c3a475ed8e38dad45a79c