RTC: drop automation call
Since the latest refactorings in the agent, there are a lot more
sections, that can be run as real-time sections.

Either way, there is no point in attaching this information to a section
plugin, since whether any given "section" is a valid choice is only
known to the agent.

CMK-6742

Change-Id: If3c6a5fee161a2632a0306bee4095530b573aa34