Move ExitSpec and state_markers out of base
None of these require base and they create unnecessary dependencies
on base on the users of these types.  This is particularly visible
on the checkers that are virtually independent on base but cannot be
moved down to the `cmk.fetchers` layer because of such types defined
in the wrong layer.

Change-Id: Iea40e8ec83ec5ebede3b5cccda588bc6f97816e3