Agent: Do not magically spawn values from config
Relying on side effects from global values is confusing and obscure the
real purpose of the functions and classes.  For example here, passing
`agent_min_version` together with `agent_target_version` explicits
an important part of the functionality of a summarizer.  Passing
a `simulation` flag is also consistent with other helpers where this
is used.

This further severes artificial dependencies and makes
the code testable.

Change-Id: Ibd31de872901257aeeb04502b4dd00127a037e2e