11857 FIX Fix changing refresh interval of views
In previous 2.0.0 releases it was not possible to use the display options
to change the refresh interval of views.

---

The display options are an implicit dependency of the painter options.

These display options were previously initialized too late from the
request variables. This has to happen before the painter options are
processed.

Change-Id: Ic5477d1e5e2f9af2f77438cdc46b5165ba729369