Regression: Fix tag lookup for piggyback host tag
The tag is now (2.0) under `HostConfig.tag_groups` instead of
`HostConfig.tag` but the code had not been adapted to reflect
this change.  In order to avoid such regressions in the future,
we call the corresponding `HostConfig` property instead of
performing a text-based lookup into the low-level tag set.

SUP-4378

Change-Id: I556e31d9dc3e2d6ad0e68c9284892fefa62a7ed4