Introduce redis for in memory caching
Especially in the GUI we often have the need to cache information for
faster access during page processing. Since the web GUI is using a
process pool model each of the pool process builds it's own cache and
consumes memory. With a larger worker pool of e.g. 50 processes, even
smaller caches can consume a significant amount of memory.

To extend our options we now introduce redis as caching daemon in the
site. It is just another daemon process that is executed in your site
and managed by your site. It will not persist any information for now.

Change-Id: I7103fbc27d66d92d5cc93b2ad84d1bc10befa7a7