11719 FIX Levels for packet rates for network interface checks
Previously, the error, multicast and broadcast WARN/CRIT levels
specified within the ruleset "Network interfaces and switch ports" were not
applied correctly. I.e., all levels were interpreted as percentual
levels, regardless of whether they had been specified as absolute or percentual
levels.
This behavior is fixed with this werk.

Additionally, the forms for error, multicast and broadcast levels now
yield a unified structure, that allows the specification of independent levels
for the in and out directions.

CMK-6472

Change-Id: I49e53098f4e612a1597d1937709a65f49ead6932