Use redis 6.2-rc1 to fix SLES12 compilation
Redis 6.0 relied on C11 features that our gcc 10.1 compiler
is not able to handle. At least with the SLES 12 distros the
compilation is not possible.

The redis team has decided to step back from the C11 requirement
which makes redis 6.2 compatible with our compiler. To make it
work on all supported distros, we now use this pre-release and
will go to 6.2 as soon as possible.

Details can be found here:

https://github.com/redis/redis/issues/6286
https://github.com/redis/redis/issues/7509

Change-Id: I0221568301fc0849966443747687955d0013f72d