Build without stashing and unstashing
The exchange of workspace packages with nexus takes a lot of
build time. Especially the parallel downloads from the nexus
take a significant build time.

Since checking out and the actual build step are performed on
the same build node, there is no need to involve remote systems.
So we copy the files around locally to reduce the time and
ressources needed.

Change-Id: Ia663d609dee3663d16e9fd1cfcb692ff5ef827b7