Serialize the packed config with pickle
The data is now serialized only once using pickle and the clear text
version is not saved anymore.  Replacing the double serialization (to
string as well as repr/eval/marshal) represents a factor 35+ speed
improvement on a small test set.  We expect this factor to scale well
to larger test sets as well.

CMK-6846

Change-Id: Icbeceadb3da63e58f10130dd4cb03c844d90a7a9