11860 SEC Password dialogs now handle passwords more securely
Previous Checkmk versions did not handle passwords well in some edit dialogs.
Especially in the setup (WATO) it was possible to access saved passwords in
some places, which should not be possible.

The dialogs now ensure that these accesses from passwords entered once and
stored on the checkmk server are no longer easily possible.

CMK-6788

Change-Id: Ibab249db3660a6b8aa140278db0a81a2269b77c0