Update utils.py
# Fixed Range Check

I realized that SIGNL4 notifications fail in beta 3 with "Details for Status Code are not defined -- 201: Created". However, 201 should be success.

The reason seems to be that the line:
if status_code in status_code_range:
returns true for 200 and 299. But the line
if status_code_range[0] <= status_code <= status_code_range[1]:
returns true for any value between 200 and 299.

The change should make sure that all 2xx status codes are processed as success.

I hope this makes sense and I did not miss anything.