11397 FIX Fix Service discovery when service checkoutput has values of -inf
This extends werk 7469, by making sure checkoutputs with values of -inf
don't crash the service discovery page.

Perfdata is unnecessary and can contain unparsable values, like "inf". If
it is passed back and forth and unparsable values appear it lead to value
errors. Simply discard perfdata when it will be used in the service
discovery page.

Extend fix of CMK-2147 and werk 7469

SUP-5026

Change-Id: Ib7d64a0b801056f1c1fb91b18a9d15e6dbba45da