Properly handle missing data in piggyback summarizer
The 'missing data' error was triggered when the piggyback file was
missing either from the directory with the hostname or from the
directory with the IP address.

The error should only be triggered when *both* are missing.

This patch fixes the logic and adds tests for this very case.

CMK-6842

Change-Id: I92f5d1fe98777002dec80e81abfee225eb23c6ef