Performance Graphs transparent area
Area plot have now a reduced opacity of 0.3 like presented in UX project
unless color already includes its opacity(ntop werk 10630).

Fillstyle was a gradient, which is defined out of bounds, thus is
completely removed for reduced opacity fill color.

Areas are not plotted as single polygon, but as a series of
trapezoids(helps with missing data), that overlap. That overlap(-1 shift)
is dropped otherwise the area looks like a zebra.

An additional line that marks the top of the area is rendered in full
opacity.

CMK-6704

Change-Id: I672bd024814acd26f987f883787e9801872e8c48