PRO TIP: make check parameters a dictionary
In the new check API the parameters of the check
funtion must be a dictionary.

If you are migrating a plugin that uses (old-style)
tuple or list parameters, you must implement a
transformation function.
It must be able to read old parameters that may
already be configured, and convert them to the new
dictionary format.

Change-Id: I764159bbc140e495a904ce844311f736115af4b0