PRO TIP: Transforms are applied during `cmk-update-config`
Since Checkmk 2.0 the transformation of check parameters
defined in WATO rules is now applied during `cmk-update-config`.

As long as a transformation is present in the corresponding
WATO rule the check plugin no longer needs to handle the
previous format.

In this case, the transformation can be found in
`cmk/gui/plugins/wato/check_parameters/mcafee_av_client.py`.

Change-Id: I55cdcf9d32252f8335adc2d7eb5f0efd328d3837