Parser FSM: Move FSM to base class
Now that we have identified the triggers and they can occur in any
state, we reduce code duplication by moving the FSM to the `__call__()`
of the base class and let the concrete state implement the `entry` and
`do` actions of each state as well as the transitions.

This requires an extra state parsing the sections of the piggybacked
data that we introduce here as well.

Change-Id: I9db70306bbe5be3462e22062a3c6cc74369f0468