First step towards uniform headers in tables
Views now have uniform group headers. If the header represents a
state its color right now spans the whole line. This will be solved
in a separate commit.

Headers for individual columns and tables in the setup menu are not
covered yet and will follow in a later commit.

CMK-6174

Change-Id: I924d2db5b7a20832caad07a16b4acba1ee9e6dff