10320 Enable SAML athentication
It is now possible to configure SAML authentication at the site apache level using mod_auth_melon.
How to do this exactly will be explained separately in a user manual article.

This feature is not supported in sles-12sp3, sles-12sp4 and sles-15 (missing dependencies).
To use this feature please upgrade to a newer Distro.

Change-Id: I4f51ae89d2e497a57f7967e3f29b91ff206be53a