checking: clean up function names
Apart from the fact that the names are changed,
there is only one small change:
The former "legacy_determine_check_params" function
actually only processed objects of the type "TimespecificParamList".
Everything else was spat out directly. In these cases
(where it does nothing), the function is now not even called.
As a result, the typing is also somewhat more precise.

Change-Id: I9ac049cacdcdf80d8e2731b89dd14794a49a76b5