Change text autocompletions by select2dropdowns
Text autocompletion was originally designed only for monitored hostnames
and configured hostnames. The text field allowed users to enter any
value and also to edit configured values. Additionally pasting a string
with particular characters(independent of autocomplete), would split it
in the case of listofstring valuespec and create new fields.

The text autocomplete suffered from some bugs and limitations:

- Options availability bug. Sometimes the request would block and no
  new options would be fetched. Sometimes the workaround of clearing the
  field, changing focus and returning would restart the query. Sometimes
  worse in the case of configured hostnames, it was necessary to exit the
  page.

- When request were not blocked, a request fired every onkeyup and when
  request came back it overwrote the user input.

This commit exchanges the adhoc select2 dialog that floats on top of the
html.input with a clean drop-down, managed entirely by select2 which
directly ajax calls to request data.  This allows for separated
concerns, the python valuespec does not need to inject JS to place hooks
to query the webapi. This fixes the following
- The data availability bug, changing input starts a new query. Thus
  results are always found
- Requests are triggered 250ms after User stops typing. This saves on
  requests.
- Select2 cancels previous request as soon as the user changes input and
  only updates contents from the response of last request. This solves the
  bug of previous request updating results and confusing user.

Old autocomplete valuespecs now inherit thus from dropdownchoice rather than text ascii.

- Monitored Hostname. Now has regex input validation. User custom input is permitted
- Configured Hostname. Now has regex input validation. User custom input is permitted
- Monitored Service. Now has regex input validation. User custom input is permitted
- Metric name. Now has input validation, metric name is strictly bound to known metrics

- Hostname and service filters. Render directly the dropdowns as filters
  don't yet use the valuespecs.

- Configured Hostname, can be embedded in listofstrings. List of strings is
extended to expand on text input and text paste when the underlying
datatype is a dropdown. It works as before.

CMK-6935

Change-Id: I7c0dac8f4ee6deeaba7d614804500b7e98c1b95d