11896 FIX Improve detailed service view
For auto-migrated check plugins the details view of the service could
contain lots of empty lines, which looked broken, in particular if the
check plugin had lots of non-OK results.

The summary output is now repeated in the details, which looks more resonable
and is in line with the new summary/details concept.

FEED-5499

Change-Id: Ib53cbfae36124bd585d74acbec21ec9c75970350