11995 FIX Availablity: GUI crash if timeline is displayed in table
The GUI crashed with the exception "Internal error: Tuple or struct_time argument required"
when the user chose to "Show timeline of each object directly in table".

Internal note:
Python3s time.mktime and time.strftime refuse to accept a list argument.
However this function changes `broken` in place, so we can not easily
change to tuple. More cleanups are ahead.

Change-Id: I52124cc7075944513135ae9aeca7abcb7b85c166