11835 Remove options to change foreground, background, canvas color of graphs
When adding graphs to views, reports or dashboards, users had the option
to change any of the colors mentioned above, which was not useful and has
been removed.

Graphs which already have any of these options set will continue to use the
user-defined colors. However, these custom colors will be removed when editing
and saving the graph the next time.

CMK-6703

Change-Id: I470e919b750d8baa6b2ef977bb996d86f9ffa303