11790 Remove state text in the "Status detail"
This change is only incompatible if you match for the state
(OK, WARN, ...) in the "Status detail" of the service output.

Previously, the "Status detail" ("Summary" in version 2.0.0) of
a service started with the state followed by the output of the
service. The service "CPU utilization" may e.g. have the output:

OK - Total CPU: 4.17%

The state is now ommitted, i.e. the output becomes:

Total CPU: 4.17%

If you match for the state in the "Status detail", you now have
to use the filter "Service states" instead.

CMK-6174

Change-Id: I80e07679388fb74ab6c47f221b42d5847e7fe7fd