Move figure dashlet base class to dashboard plugin utils
This class is part of the dashboard plugin domain and needs to be available
for the plugins via cmk.gui.plugins.dashboard and for the main dashboard.py
file.

Imports of dashboard code in cmk.gui.figures together with imports of cmk.gui.figures
in the dashboard code resulted in a cycle which is cleaned up with this change.

Change-Id: Ia27e47c2fa0113fb560a309a9884fc0fcbd51b4f