Parametrizable valuespec for dashlet data generators
New dashlets used a static data generator, which was responsible of
preparing necessary data based on a static valuespec. To avoid presenting
the user useless options that have no effect depending on the selected
dashlet, the dashlet needs to influence which properties of the valuespec
it wants to exclude.

This change implies that the method to process the requests can't be a
classmethod, callable without instantiating the data generator class. As
the valuespec depends on the instantiated class. The ajax data generator
endpoints just present the unrestricted data. They are in the end generic
endpoints.

Change-Id: I387e8cf83f43c8a3a89e88f3fd1fc46d60ad1ac6