11842 Stabilize the Windows agent updater
In some cases for various reasons the agent updater of the  Windows agent
can't update the agent. There are two possible outcomes: The agent is not
updated at all or the agent is even uninstalled.

Since this release, the Agent controls whether the update was successful
and on failure will retry update. The additional information is stored in
the file <tt>%ProgramData%\checkmk\agent\log\update.log</tt>

Change-Id: I7ee14ee554252526e366b6d20b93dfe2fd3cecc4