DRAFT Remove data generator class for dashlet
Currently dashlets query an ajax endpoint, which is just a page, that
replies to the dashlet data generator. These data generators come in a pair
match to dashlet and thus become to rigid to serve multiple dashlets that
only require a subset of the data or worse that can incompatible render
attributes.

Inspecting the ABCDatagenerator the only relevant method is
generate_response_from_request. Within it, it was already capable to find
the dashlet being draw, yet it extracted the configuration valuespec out of
itself(DataGenerator) instead of the Dashlet, which is the correct
candidate to know the required configuration valuespec.

This refactoring taking as example the Single metric source, that can be
displayed as a Gauge(w/o history) or Single Graph(w/o history) for a
specific host and service or a Barplot with current value for multiple
hosts. All configuration valuespec is transferred to the
dashlets. Datagenerator is removed leaving some middle ware not to break
other dashlets. And the WEB API ajax endpoint is named for single metric,
but could easily process all new figure dashlets, because the dashlet type
is recognized during processing not on API endpoint.

The specificity of the configuration valuespec at the dashlet level allows
for flexible and unique dashlet endpoint and correct data validation.

Change-Id: I15b8561f1011a44f09348cfb3005af715ca484ea