dell_{compellent_*,hw_info}: share a detection OID
This decreases the OIDs that must be fetched from *all* SNMP devices
during discovery by one.
It comes at the cost of fetching one OID more for all devices
that provide the Dell specific table ".1.3.6.1.4.1.674".

This 'hack' is used, because we currently have no relieable information
on the common system description of Dell-Compellent devices.

CMK-5889

Change-Id: I205a2581a663c893bad35a25117fefd8162d7a31