printer_*: share a detection OID
This decreases the OIDs that must be fetched from *all* SNMP devices
during discovery by one.
It comes at the cost of fetching one OID more for all printers
(devices that provide the table ".1.3.6.1.2.1.43").

This 'hack' is used, because we have no relieable information
on the common system description of all printers.

We might use a hardcoded list of printer vendors in the future.

CMK-5889

Change-Id: I68af606c797eb1c608f579895d6316c3f9fb0b31