inv_if,if32: share a detection OID
The section "inv_if" used to be detected, if a device had at least two (!)
network interfaces. It will now be detected, if an ifTable is present.

Note that this section is currently only used in a HaSI plugin, so this
change has no practical effect (apart from fetching one OID less during discovery).

Change-Id: I086b94f0e625a9e0c3ed7c6c3fd75037d5584096