11722 FIX Bug occuring on xinetd reload on agent update
On (automatic) agent update on Linux, the update process includes the reload of
the xinetd process, if running. This is done by the DEB maintainer scripts or
RPM scriptlets, respectively.

There are some rare situations (e.g., LCX containers with a running xinetd) that lead
to a falsely detected xinetd process on the host system. In that case, the update
process would crash on trying to reload xinetd.

This behavior is mitigated with this werk, by only detecting running xinetd processes
that run under root (uid 0).

CMK-6617

Change-Id: Iccb4d68a55906fd4c4d60324cc79d3156f9e7b7b