12031 FIX Views: Views are now asking for missing contexts just like dashboards
When a view that requires a context (e.g. as host name) is opened without
a "host" HTTP request variable, the view only showed an error message and
stopped rendering the view. The user then had to add the missing context
information to the view configuration or to the URL variables.

From now on, the view is still displaying a warning message like "Unable to
render this view, because we miss some required context information (host).
Please update the form on the right to make this view render.", but also
displaying the filter form asking the user to provide the missing context.

Change-Id: I2338f9645629c4934a3d37e44b8e565b42434358