PRO TIP: make a plugin fit for clustering
In the old API, all plugins could be applied on a cluster.
They would do something more or less usefull, depending on
whether the developer had clustering in mind when designing
the parse and check functions. In the new API, a plugin must
implement the cluster mode explicitly.

Previously, the received monitoring data was concatenated
(in the same order the nodes where assigned to the cluster)
and the resulting "cluster data" was passed to the parse
function.

In the new API the parse function is always applied to data
corresponding to one single node. The resulting parsed
sections are collected into one dictionary, with the nodes
name as key.

These "clustered sections" are then passed to the cluster
check function, that will be called *instead* of the check
function if the service is clustered.

In this case we are bringing back the cluster capability of
a plugin that has been migrated from the old API -- so we
wish to implement the old behaviour.

We combine the separately parsed datasets (only for the
relevant item) in a way that results in the same dataset that
we would have gotten in the old API, by firstly concatenating
the lines, and parsing them together.

Change-Id: Icbb61055f0f0a1e025051a7b0527d58590ceeae2