11901 FIX Fix crash in ESX Inventory
The format of the data was changed in WERK 8143, which could possibly
result in crashes.

This partly reverts to a state before
5f3309f3e0ee6025f717d5424d6e58826e26509f

It is not clear where the new hierarchical format comes from but the
patch has been responsible for the crashes:

  https://crash.checkmk.com/gui/crashreportgroupview/show/1810

This patch lets the code handle the legacy (list-based) format as well
as the new (dict-based) one.

Change-Id: I387064121b84d8bf095bede0dcfb9a807332e7a6