discovery: rename for clarity
We have way to many variants of `do_foo`, `execute_foo`
and `perform_foo` variants of function names.

We try to bring a *bit* of clarity.
In this commit:

  * more usage of kwargs
  * hostname -> host_name
  * _perform_host_label_discovery -> _analyse_host_labels
  * _execute_discovery -> _discover_plugins_services

Ultimately, we would like to find a new name for the
process of integrating and comparing the existing
services with the discovered ones.

Currenlty this is refered to as 'performing' the discovery,
which is confusing.

'Discovery' is the process of extracting the services or
host labels from the present monitoring data.

Change-Id: Id7d1d184229744c0485f39fa3388a8488213b28c