11352 FIX iptables: Keep ACK after update/restart
In case a changed iptable config - which results in {CRIT} - was acknowledged, the ACK vanished after a <tt>omd update<\tt> or <tt>omd restart<\tt>.
The reason therfore is that the check will go to state {OK} for one checkinterval after the update/restart and then become {CRIT} again.
The intermediate state is now {PENDING}, which is actually more precise.

Change-Id: If79680696ea85c36d894d9b5babba09e41312a8c