Improve IPMI error message
An error during the IPMI connection (such as a timeout) results in a
traceback with a low-level error message or an error message that just
says "None".  (See code in pyghmi.ipmi.private.session._mark_broken()).

We replace the IpmiException with our MKFetcherError and add a slightly
more useful error message.

CMK-7011

Change-Id: Iad40210e40eceafd65d138e947c536611521be94