12038 FIX Fix displaying host graphs on dashboard
Previous 2.0.0 versions could not display host related time series graphs on
the dashboard. The new context validation mechanism, that expects a service
context for this dashlet, was preventing the rendering of that dashlet. Now
the host dashlets automatically get an empty service context set which makes
the dashlet render again.

Change-Id: I8a443775117dced2a878f757aa2c8693409df750