12236 <tt>oracle_instance</tt>: More detailed version information
The check plugin <tt>oracle_instance</tt> (services "ORA ... Instance"),
which monitors the state of Oracle instances, now displayes more detailed
version information if the Oracle version of the monitored host supports
this (the minimum required version for this is Oracle 18c).

For this change to come into effect, you need to update the Checkmk agent
monitoring your Oracle hosts. The displayed version will for example change
from Version 19.0.0.0.0 to Version 19.5.0.0.0. This has no impact on the
monitoring state of the corresponding services.

Closes #323

Change-Id: I78f7da9b0a924da72cba8bf8033c199256aecb0d