11726 FIX Error on windows agent update in case of a numerical agent hash
After an (automatic) windows agent update took place,
the agent updater checks for a correct agent hash on the next call.

Previously, the check failed, if the new hash only contains numbers, with
the acutal update being successful.
As a result, the agent updater downloaded the agent periodically, while the
Check_MK Agent service periodically changed to WARN with a last_error of
<em>inconsistent pending update</em>.

With this Werk applied, the behavior will be fixed after the next agent baking
and the succeeding agent update on the affected hosts.

Change-Id: I053374d5d2c730e28aad39d1c8e8fcc2cf06805b