11846 Windows agent supports logging for winperf
Since this release it is possible to log messages from winperf.

To get log you have to edit check_mk.user.yml file and set in the
section <tt>winperf</tt> the key <tt>trace</tt> to <tt>yes</tt>.

Restart of the service is mandatory. The log will be provided in the file
<log/winperf.log</tt>

Change-Id: I3193b72ea1a6cd1544e5370aebe8d1f23b76eb82