10133 FIX Fixed source attribute of user comments
The CMC always reported all comments as coming from an "external" source,
even acknowledgements.  Nagios considers the latter as "internal", so we do
the same now for consistency.

There is no impact on Checkmk's GUI, it doesn't use the "source" column from
the "comments" table.  You can only observe the change directly via
Livestatus if you use that column, which is highly unlikely.

Change-Id: I1ccafaadbdb90a8b78485dbd4de3e16aa4eff61f