discovery: consolidate caching options
The handling of the caching options is consolidated into
the discovery module. The code in cmk.base.automations only
passes flags along.

The functions
 * `discover_on_host`
    used by the automation 'inventory' and the automatic
    service discovery and
 * `get_check_preview`
    used by the automation 'try-inventory'
no longer accept the argument 'use_caches'. Instead, they
are passed the arguments 'use_cached_snmp_data' and
'max_cachefile_age'.

The flag 'use_caches' was aplied to both the TCP and the
SNMP data sources. That is wrong:

The TCP data sources could be sending some data only once,
in which case information would be missed by the checks.
TCP sources *must* use the caches.

The cache file of the SNMP data source will only contain
the sections detected when the data was fetched. The set of
detected sections is expected to change rarely, so
depending on the use case, using the SNMP caches may not be
desired. In particular: When hitting the "Full scan" button
on the discovery page, we expect new sections to be
detected.

Change-Id: I4b801a382e62ebdfa8d08788cdf28580a5ac5397