Simplify integration test setup
As very first step of the site setup the GUI language of the testing
user is set to english to execute all tests in a non localized
environment.

This setting of the language was done by submitting a POST request
simulating the "edit form" dialog being filled in.  This is something
which was very error prone in the past.

In order to eliminate the source of errors once and for all, this change
is now made via the Web API (will be converted to new REST API soon).

Change-Id: I36f0a347545f3727a83339a5018f3b969b8cc102