12043 FIX Service discovery: Disabling services works also with disabled rule
The service discovery page created and updates "disabled services" rules
in the moment a user is clicking on the "disable this service" icon on
the service discovery page.

In the moment a user disabled such a rule, the "disable this service" feature
on the service discovery page did not work anymore because the related rule
was disabled.

From now on the disabled rules are not updaten anymore in such a situation.
Instead the service discovery page will create a new rule to make the disabling
of the service work.

Change-Id: I71b4630acf3a493c9477d88406ec043496cad10d